refactor(orchestrator): clean up orchestrator process by creating helpers methods for spawn/verify/introspector/span_collator
