+++ javascript/packages/orchestrator/src/providers/index.ts
@@ -6,3 +6,14 @@ export const Providers = new Map();
 Providers.set("kubernetes", k8sProvider);
 Providers.set("podman", podmanProvider);
 Providers.set("native", nativeProvider);
+
+export function getProvider(provider: string) {
+  if (!Providers.has(provider)) {
+    throw new Error(
+      "Invalid provider config. You must one of: " +
+        Array.from(Providers.keys()).join(", "),
+    );
+  }
+
+  return Providers.get(provider);
+}
