+++ javascript/packages/orchestrator/src/types.ts
@@ -244,6 +244,8 @@ export interface ChainSpec {
   };
 }
 
+export type NodeMultiAddress = string;
+
 // Utils
 export interface GlobalVolume {
   name: string;
